Rename everything
