--- collision_msgs/msg/Collision.msg
@@ -1,2 +1,0 @@
-string entity0
-string entity1
