--- collision_msgs/msg/Collisions.msg
@@ -1,2 +1,0 @@
-std_msgs/Header header
-collision_msgs/Collision[] collisions
