+++ collision_msgs/msg/NamedCollision.msg
@@ -1,0 +1,4 @@
+# Message to indicate a collision between two entities.
+# The second entity name can be blank if a sensor detects a collision but doesn't have a name for the object
+string entity0
+string entity1
