+++ collision_msgs/msg/NamedCollisions.msg
@@ -1,0 +1,2 @@
+std_msgs/Header header
+collision_msgs/NamedCollision[] collisions
