+++ compute_benchmark_msgs/CHANGELOG.rst
@@ -1,0 +1,8 @@
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package compute_benchmark_msgs
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+0.1.0 (2023-08-29)
+------------------
+* benchmark_msgs, benchmark_utils, collision_msgs
+* Contributors: David V. Lu!!
