+++ compute_benchmarking/include/compute_benchmarking/benchmark_publisher.hpp
@@ -34,63 +34,40 @@
 
 /* Author: David V. Lu!! */
 
-#include <benchmark_utils/benchmark_publisher.hpp>
+#pragma once
 
-namespace benchmark_utils
+#include <rclcpp/rclcpp.hpp>
+#include <compute_benchmark_msgs/msg/compute_time.hpp>
+
+#include <chrono>
+#include <string>
+
+namespace compute_benchmarking
 {
-const int NANOMOD = 1000000000;
-
-BenchmarkPublisher::BenchmarkPublisher(const rclcpp::Node::SharedPtr& node, const std::string& topic) : node_(node)
+typedef struct BenchmarkContextS
 {
-  if (node_)
-    pub_ = node_->create_publisher<benchmark_msgs::msg::ComputeTime>(topic, 1);
-}
+  std::string name;
+  std::string id;
+  std::chrono::high_resolution_clock::time_point start;
 
-void BenchmarkPublisher::tick(const std::string& name)
+  BenchmarkContextS(const std::string& name, const std::string& id)
+    : name(name), id(id), start(std::chrono::high_resolution_clock::now())
+  {
+  }
+} BenchmarkContext;
+
+class BenchmarkPublisher
 {
-  int c;
-  if (counter_.count(name) == 0)
-  {
-    c = 0;
-  }
-  else
-  {
-    c = counter_[name];
-  }
-  counter_[name] = c + 1;
-  std::string ident = name + "_" + std::to_string(c);
-  stack_.emplace_back(BenchmarkContext(name, ident));
-}
+public:
+  BenchmarkPublisher(const rclcpp::Node::SharedPtr& node, const std::string& topic);
 
-double BenchmarkPublisher::tock(bool log)
-{
-  std::chrono::high_resolution_clock::time_point stop = std::chrono::high_resolution_clock::now();
-  BenchmarkContext& cxt = stack_.back();
-  std::chrono::nanoseconds duration_nano = stop - cxt.start;
-  long d_nano_i = duration_nano.count();
-  long s_nano_i = stop.time_since_epoch().count();
+  void tick(const std::string& name);
+  double tock(bool log = true);
 
-  benchmark_msgs::msg::ComputeTime msg;
-  msg.header.frame_id = cxt.name;
-  msg.header.stamp.sec = s_nano_i / NANOMOD;
-  msg.header.stamp.nanosec = s_nano_i % NANOMOD;
-  msg.duration.sec = d_nano_i / NANOMOD;
-  msg.duration.nanosec = d_nano_i % NANOMOD;
-  msg.id = cxt.id;
-
-  stack_.pop_back();
-  if (!stack_.empty())
-  {
-    msg.parent_id = stack_.back().id;
-  }
-  if (pub_)
-    pub_->publish(msg);
-  double duration_sec_f = d_nano_i / 1e9;
-  if (log && node_)
-  {
-    RCLCPP_INFO(node_->get_logger(), "%s time: %.4f", msg.header.frame_id.c_str(), duration_sec_f);
-  }
-
-  return duration_sec_f;
-}
-}  // namespace benchmark_utils
+protected:
+  std::list<BenchmarkContext> stack_;
+  std::unordered_map<std::string, int> counter_;
+  rclcpp::Node::SharedPtr node_;
+  rclcpp::Publisher<compute_benchmark_msgs::msg::ComputeTime>::SharedPtr pub_;
+};
+}  // namespace compute_benchmarking
